user: fix compilation on 32-bit platforms

+++ user/libs/conf.c
@@ -415,12 +415,16 @@ rl_conf_flows_print(struct list_head *flows)
 
         PI_S("  ipcp %u, local addr/port %llu:%u, "
              "remote addr/port %llu:%u, %s"
-             "tx %lu pkt %lu byte %lu err, "
-             "rx %lu pkt %lu byte %lu err\n",
-             rl_flow->ipcp_id, (long long unsigned int)rl_flow->local_addr,
-             rl_flow->local_port, (long long unsigned int)rl_flow->remote_addr,
-             rl_flow->remote_port, specinfo, stats.tx_pkt, stats.tx_byte,
-             stats.tx_err, stats.rx_pkt, stats.rx_byte, stats.rx_err);
+             "tx %llu pkt %llu byte %llu err, "
+             "rx %llu pkt %llu byte %llu err\n",
+	     rl_flow->ipcp_id, (long long unsigned int)rl_flow->local_addr,
+	     rl_flow->local_port, (long long unsigned int)rl_flow->remote_addr,
+	     rl_flow->remote_port, specinfo, (long long unsigned)stats.tx_pkt,
+	     (long long unsigned)stats.tx_byte,
+	     (long long unsigned)stats.tx_err,
+	     (long long unsigned)stats.rx_pkt,
+	     (long long unsigned)stats.rx_byte,
+	     (long long unsigned)stats.rx_err);
     }
 
     return 0;
